# doc/api.yaml
openapi: 3.0.0
servers:
  - url: http://localhost:3000/api/v3
info:
  title: WASAText
  description: Web And Software Architecture Exam Project
  version: 1.0.0
security:
  - bearerAuth: []
...
paths:
...
  /messages/{messageId}/receipts:
    parameters:
      - name: messageId
        in: path
        description: Message id
        required: true
        schema:
          type: integer
          format: int64
    put:
      tags: ["messages"]
      summary: Set the message status to 'read'
      operationId: setMessageStatusAsRead
      responses:
        '204':
          description: No Content
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
    get:
      tags: ["messages"]
      summary: Get receipts of the specified message id
      description: Get receipts of the specified message id
      operationId: getReceipts
      responses:
        '200':
          description: Successful
          content:
            application/json:
              example:
                type: object
              schema:
                description: Receipts
                type: object
                properties:
                  receipts:
                    description: Receipts
                    type: array
                    minItems: 1
                    maxItems: 1024
                    items:
                      $ref: "#/components/schemas/Receipt"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
  /users:
    get:
      tags: ["users"]
      summary: Get WASAText users
      description: Get WASAText users
      operationId: getUsers
      responses:
        '200':
          description: Successful
          content:
            application/json:
              example:
                type: object
              schema:
                description: Users of WASAText
                type: object
                properties:
                  users:
                    description: Users
                    type: array
                    minItems: 1
                    maxItems: 1024
                    items:
                      $ref: "#/components/schemas/User"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        description: User id
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: ["users"]
      summary: Get user data
      description: Get user data
      operationId: getUserById
      responses:
        '200':
          description: Successful
          content:
            application/json:
              example:
                type: object
              schema:
                $ref: "#/components/schemas/User"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: ["users"]
      summary: Delete user
      description: Delete user
      operationId: deleteUser
      responses:
        '204':
          description: No Content
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/name:
    parameters:
      - name: userId
        in: path
        description: User id
        required: true
        schema:
          type: integer
          format: int64
    put:
      tags: ["users"]
      operationId: setMyUserName
      summary: Update user name
      description: Update user name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: The username
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/Name"
      responses:
        '200':
          description: Successful
          content:
            application/json:
              example:
                type: object
              schema:
                description: The username
                type: object
                properties:
                  name:
                    $ref: "#/components/schemas/Name"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '409':
          $ref: "#/components/responses/Conflict"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/photo:
    parameters:
      - name: userId
        in: path
        description: User id
        required: true
        schema:
          type: integer
          format: int64
    put:
      tags: ["users"]
      operationId: setMyPhoto
      summary: Update user photo
      description: Update user photo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              description: The user photo
              type: object
              properties:
                photo:
                  $ref: "#/components/schemas/Photo"
      responses:
        '200':
          description: Successful
          content:
            application/json:
              example:
                type: object
              schema:
                description: The photo URL
                type: object
                properties:
                  photoUrl:
                    $ref: "#/components/schemas/PhotoUrl"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
  /session:
    post:
      tags: ["authentication"]
      summary: Logs in the user
      description: Logs in the user
      operationId: doLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Request body
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/Name"
      responses:
        '201':
          description: Created
          content:
            application/json:
              example:
                type: object
              schema:
                description: Created
                type: object
                properties:
                  userId:
                    description: User id
                    type: integer
                    format: int64
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
  schemas:
    User:
      description: User
      type: object
      properties:
        userId:
          description: User id
          type: integer
          format: int64
        name:
          $ref: "#/components/schemas/Name"
        photoUrl:
          $ref: "#/components/schemas/PhotoUrl"
...
    Name:
      description: Generic name
      type: string
      pattern: '^.*$'
      minLength: 1
      maxLength: 1024
    PhotoUrl:
      description: Photo URL
      type: string
      format: url
      minLength: 1
      maxLength: 1024
...
    Receipt:
      description: Receipt
      type: object
      properties:
        messageId:
          description: Message id
          type: integer
          format: int64
        userId:
          description: User id
          type: integer
          format: int64
        status:
          description: Status
          type: string
          enum: [sent, received, read]
        sentAt:
          description: Sent at
          type: string
          format: date-time
        receivedAt:
          description: Received at
          type: string
          format: date-time
        readAt:
          description: Read at
          type: string
          format: date-time
...
    Photo:
      description: Photo
      type: string
      format: binary
      minLength: 1
      maxLength: 1014
...