# service/api/users.go
package api

import (
	"database/sql"
	"errors"
	"fmt"
	"net/http"
	"strconv"

	"github.com/gensimone/WASA-project/service/database"
	"github.com/julienschmidt/httprouter"
)

// operationId: getUsers
func (rt *_router) getUsers(w http.ResponseWriter, _ *http.Request, _ httprouter.Params, _ database.User) {
	userIds, err := rt.db.GetUserIds()
	if err != nil {
		rt.sendResponse(w, "Internal Sever Error", http.StatusInternalServerError)
		rt.baseLogger.Errorf("GetUserIds: %v", err)
		return
	}

	rt.sendResponse(w, struct {
		UserIds []int64 `json:"users"`
	}{UserIds: userIds}, http.StatusOK)
}

// operationId: getUserById
func (rt *_router) getUserById(w http.ResponseWriter, _ *http.Request, ps httprouter.Params, _ database.User) {
	userId, err := strconv.ParseInt(ps.ByName("userId"), 10, 64)
	if err != nil {
		rt.sendResponse(w, "Parameter userId must be an int64", http.StatusBadRequest)
		return
	}

	user, err := rt.db.GetUserById(userId)

	switch {
	case errors.Is(err, sql.ErrNoRows):
		rt.sendResponse(w, fmt.Sprintf("User id %d not found", userId), http.StatusNotFound)
	case err != nil:
		rt.sendResponse(w, "Internal Server Error", http.StatusInternalServerError)
		rt.baseLogger.Errorf("GetUserById: %v", err)
	default:
		rt.sendResponse(w, user, http.StatusOK)
	}
}

// operationId: deleteUser
func (rt *_router) deleteUser(w http.ResponseWriter, _ *http.Request, ps httprouter.Params, user database.User) {
	userId, err := strconv.ParseInt(ps.ByName("userId"), 10, 64)
	if err != nil {
		rt.sendResponse(w, "Parameter userId must be an int64", http.StatusBadRequest)
		return
	}

	if userId != user.UserId {
		rt.sendResponse(
			w,
			"Parameter userId must be equal to the provided authentication id",
			http.StatusBadRequest,
		)
		return
	}
	_, err = rt.db.DeleteUser(user.UserId)
	if err != nil {
		rt.sendResponse(w, "Internal Server Error", http.StatusInternalServerError)
		rt.baseLogger.Errorf("DeleteUser: %v", err)
		return
	}

	rt.sendResponse(w, nil, http.StatusNoContent)
}

// operationId: setMyUserName
func (rt *_router) setMyUserName(w http.ResponseWriter, r *http.Request, ps httprouter.Params, user database.User) {
	userId, err := strconv.ParseInt(ps.ByName("userId"), 10, 64)
	if err != nil {
		rt.sendResponse(w, "Parameter userId must be an int64", http.StatusBadRequest)
		return
	}

	if userId != user.UserId {
		rt.sendResponse(
			w,
			"Parameter userId must be equal to the provided authentication id",
			http.StatusBadRequest,
		)
		return
	}

	name, err := rt.getNameFromReq(w, r)
	if err != nil {
		return
	}

	_, err = rt.db.GetUserByName(*name)

	switch {
	case errors.Is(err, sql.ErrNoRows):
		_, err := rt.db.SetMyUserName(user.UserId, *name)
		if err != nil {
			rt.sendResponse(w, "Internal Server Error", http.StatusInternalServerError)
			rt.baseLogger.Errorf("SetMyUserName: %v", err)
			return
		}

		rt.sendResponse(w, struct {
			Name string `json:"name"`
		}{Name: *name}, http.StatusOK)
	case err != nil:
		rt.sendResponse(w, "Internal Server Error", http.StatusInternalServerError)
		rt.baseLogger.Errorf("GetUserByName: %v", err)
	default:
		rt.sendResponse(w, "User name already taken", http.StatusConflict)
	}
}

// operationId: setMyPhoto
func (rt *_router) setMyPhoto(w http.ResponseWriter, r *http.Request, ps httprouter.Params, user database.User) {
	userId, err := strconv.ParseInt(ps.ByName("userId"), 10, 64)
	if err != nil {
		rt.sendResponse(w, "Parameter userId must be an int64", http.StatusBadRequest)
		return
	}

	if userId != user.UserId {
		rt.sendResponse(
			w,
			"Parameter userId must be equal to the provided authentication id",
			http.StatusBadRequest,
		)
		return
	}

	photoUrl, err := rt.uploadMediaFile(w, r, "photo")
	if err != nil {
		return
	}

	if _, err := rt.db.SetMyPhotoUrl(user.UserId, photoUrl); err != nil {
		rt.baseLogger.Errorf("SetMyPhotoUrl: %v", err)
		_ = rt.removeMediaFile(photoUrl)
		rt.sendResponse(w, "Internal Server Error", http.StatusInternalServerError)
		return
	}

	// NOTE: Here we must delete the old photo.
	//       Or maybe we could remove old photos (or even attachments) somewhere else (e.g. a script)
	rt.sendResponse(w, struct {
		PhotoUrl string `json:"photoUrl"`
	}{PhotoUrl: photoUrl}, http.StatusOK)
}

// operationId: doLogin
func (rt *_router) doLogin(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
	name, err := rt.getNameFromReq(w, r)
	if err != nil {
		return
	}

	user, err := rt.db.GetUserByName(*name)
	var userId *int64
	switch {
	case errors.Is(err, sql.ErrNoRows):
		userId, err = rt.db.CreateUser(*name, rt.defaultUserPhoto)
		if err != nil {
			rt.sendResponse(w, "Internal Server Error", http.StatusInternalServerError)
			rt.baseLogger.Errorf("CreateUser: %v", err)
			return
		}
	case err != nil:
		rt.sendResponse(w, "Internal Server Error", http.StatusInternalServerError)
		rt.baseLogger.Errorf("GetUserByName: %v", err)
		return
	default:
		userId = &user.UserId
	}

	rt.sendResponse(w, struct {
		UserId int64 `json:"userId"`
	}{UserId: *userId}, http.StatusCreated)
}